Refactor graphql currency queries
