+++ server/graphql/general/src/lib.rs
@@ -9,7 +9,10 @@ use chrono::{DateTime, Utc};
 use graphql_core::pagination::PaginationInput;
 
 use crate::store_preference::store_preferences;
-use graphql_types::types::{StorePreferenceNode, TemperatureLogFilterInput};
+use graphql_types::types::{
+    CurrenciesResponse, CurrencyResponse, CurrencySortInput, StorePreferenceNode,
+    TemperatureLogFilterInput,
+};
 use mutations::{
     barcode::{insert_barcode, BarcodeInput},
     common::SyncSettingsInput,
@@ -23,6 +26,7 @@ use mutations::{
     update_user,
 };
 use queries::{
+    currency::{get_currencies, get_currency},
     display_settings::{display_settings, DisplaySettingsHash, DisplaySettingsNode},
     initialisation_status::{initialisation_status, InitialisationStatusNode},
     requisition_line_chart::{ConsumptionOptionsInput, StockEvolutionOptionsInput},
@@ -294,6 +298,23 @@ impl GeneralQueries {
             number_of_data_points,
             filter,
         )
+    }
+
+    pub async fn currency(
+        &self,
+        ctx: &Context<'_>,
+        currency_id: String,
+    ) -> Result<CurrencyResponse> {
+        get_currency(ctx, &currency_id)
+    }
+
+    pub async fn currencies(
+        &self,
+        ctx: &Context<'_>,
+        #[graphql(desc = "Sort options (only first sort input is evaluated for this endpoint)")]
+        sort: Option<Vec<CurrencySortInput>>,
+    ) -> Result<CurrenciesResponse> {
+        get_currencies(ctx, sort)
     }
 }
