+++ server/graphql/general/src/queries/mod.rs
@@ -46,6 +46,8 @@ pub use self::last_successful_user_sync::*;
 pub use self::plugin::*;
 pub mod temperature_chart;
 pub use self::temperature_chart::*;
+pub mod currency;
+pub use self::currency::*;
 
 #[cfg(test)]
 mod tests;
