+++ server/graphql/lib.rs
@@ -15,7 +15,6 @@ use graphql_batch_mutations::BatchMutations;
 use graphql_clinician::ClinicianQueries;
 use graphql_core::loader::LoaderRegistry;
 use graphql_core::{auth_data_from_request, BoxedSelfRequest, RequestUserData, SelfRequest};
-use graphql_currency::CurrencyQueries;
 use graphql_form_schema::{FormSchemaMutations, FormSchemaQueries};
 use graphql_general::{
     DiscoveryQueries, GeneralMutations, GeneralQueries, InitialisationMutations,
@@ -73,7 +72,6 @@ pub struct Queries(
     pub FormSchemaQueries,
     pub ClinicianQueries,
     pub PluginQueries,
-    pub CurrencyQueries,
 );
 
 impl Queries {
@@ -97,7 +95,6 @@ impl Queries {
             FormSchemaQueries,
             ClinicianQueries,
             PluginQueries,
-            CurrencyQueries,
         )
     }
 }
